+++ pipeline/chat.h
@@ -13,7 +13,7 @@ public:
     ChatEngine();
     ~ChatEngine();
 
-    bool init(const std::string& model_path_, DevMode dev_mode = DevMode::CPU_MODE, int n_threads = 1, int max_seq_len = 2048, float top_p = -1.0f, int top_k = -1, float temperature = -1.0f, int gpu_layer = 0);
+    bool init(const std::string& model_path_, DevMode dev_mode = DevMode::CPU_MODE, int n_threads = 1, int max_seq_len = 2048, float top_p = -1.0f, int top_k = -1, float temperature = -1.0f, size_t gpu_layer = 0);
     bool run_complete(const std::string& prompt, const int max_tokens, std::string& response);
 
 private:
